GUI Refactor.
