+++ gui/index.html
@@ -6,8 +6,8 @@
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=League+Spartan:wght@200;300;600&display=swap" rel="stylesheet"> 
-    <script src="/wails/ipc.js"></script>
-    <script src="/wails/runtime.js"></script>
+    <!--<script src="/wails/ipc.js" type="module"></script>-->
+    <script src="/wailsjs/runtime/runtime.js" type="module"></script>
     <title>CFN Tracker v2 by @greensoap_</title>
   </head>
   <body>
