+++ gui/package.json
@@ -9,24 +9,29 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@xstate/react": "^3.2.2",
     "i18next": "^22.4.14",
     "i18next-browser-languagedetector": "^7.0.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-hook-form": "^7.44.2",
     "react-i18next": "^12.1.4",
     "react-icons": "^4.7.1",
     "react-minimal-pie-chart": "^8.4.0",
     "react-router-dom": "^6.6.2",
+    "xstate": "^4.37.2",
     "zustand": "^4.3.2"
   },
   "devDependencies": {
+    "@tailwindcss/forms": "^0.5.3",
+    "@types/node": "^20.2.5",
     "@types/react": "^18.0.17",
     "@types/react-dom": "^18.0.6",
     "@vitejs/plugin-react": "^2.1.0",
     "autoprefixer": "^10.4.12",
     "postcss": "^8.4.17",
+    "sass": "^1.62.1",
     "tailwindcss": "^3.1.8",
-    "@tailwindcss/forms": "^0.5.3",
     "typescript": "^4.6.4",
     "vite": "^3.1.0"
   },
