+++ gui/package.json.md5
@@ -1,1 +1,1 @@
-cc4f14ca0fc7cf74281c2c5c765a1949
+3dc30a3a49f27ed8a9235f094d057a1c
