+++ gui/src/i18n.ts
@@ -1,9 +1,21 @@
 import i18n from 'i18next';
-import { initReactI18next } from 'react-i18next';
 import LanguageDetector from 'i18next-browser-languagedetector';
+import { initReactI18next } from 'react-i18next';
+
 import eng from './locales/en.json';
 import fr from './locales/fr.json';
 import jp from './locales/jp.json';
+
+type AppLanguage = {
+  code: string;
+  nativeName: string;
+}
+
+export const APP_LANGUAGES: AppLanguage[] = [
+  { code: 'en', nativeName: 'English' },
+  { code: 'fr', nativeName: 'Français' },
+  { code: 'jp', nativeName: '日本' },
+];
 
 i18n
   .use(LanguageDetector)
