+++ gui/vite.config.ts
@@ -1,7 +1,14 @@
 import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react'
+import path from "path";
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [react()]
+  plugins: [react()],
+  resolve: {
+    alias: {
+      '@': path.resolve(__dirname, "./src/"),
+      '@@': path.resolve(__dirname, "./wailsjs/")
+    }
+  }
 })
